Move environment setting to the sign in screen
