+++ apps/ios/LiveVideo/LiveVideo/Managers/AppSettings/AppSettingsManager.swift
@@ -6,4 +6,8 @@ import SwiftUI
 
 class AppSettingsManager: ObservableObject {
     @AppStorage("TwilioEnvironment") var environment: TwilioEnvironment = .prod
+
+    func reset() {
+        UserDefaults.standard.removePersistentDomain(forName: Bundle.main.bundleIdentifier!)
+    }
 }
